requirements-dev-snowpark.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asn1crypto==1.5.1
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
backports-tarfile==1.2.0
backrefs==5.9
beautifulsoup4==4.14.3
black==24.10.0
bleach[css]==6.3.0
blinker==1.9.0
boto3==1.42.21
botocore==1.42.21
bracex==2.6
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
coverage[toml]==7.6.10
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.76.0
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
docker==7.1.0
docutils==0.21.2
exceptiongroup==1.3.1
execnet==2.1.2
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.2
flask==3.1.2
flit==3.12.0
flit-core==3.12.0
fonttools==4.61.1
fqdn==1.5.1
ghp-import==2.1.0
gitdb==4.0.12
gitpython==3.1.46
google-auth==2.45.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
griffe==1.15.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
id==1.5.0
idna==3.10
importlib-metadata==8.7.1
importlib-resources==6.5.2
iniconfig==2.3.0
ipykernel==6.31.0
ipython==8.37.0
isoduration==20.11.0
itsdangerous==2.2.0
jaraco-classes==3.4.0
jaraco-context==6.0.2
jaraco-functools==4.4.0
jedi==0.19.2
jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema[format-nongpl]==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.7.0
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.3
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupytext==1.18.1
keyring==25.7.0
kiwisolver==1.4.9
lark==1.3.1
mako==1.3.10
markdown==3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mergedeep==1.3.4
mike==2.1.3
mistune==3.2.0
mkdocs==1.6.1
mkdocs-autorefs==1.4.3
mkdocs-dracula-theme==1.0.7
mkdocs-get-deps==0.2.0
mkdocs-git-committers-plugin-2==2.5.0
mkdocs-include-markdown-plugin==7.1.5
mkdocs-jupyter==0.25.1
mkdocs-material==9.6.14
mkdocs-material-extensions==1.3.1
mkdocstrings[python]==0.29.1
mkdocstrings-python==1.16.12
mlflow==2.20.1
mlflow-skinny==2.20.1
more-itertools==10.8.0
mypy-extensions==1.1.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.3.2
notebook-shim==0.2.4
numpy==1.23.5
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
overrides==7.7.0
packaging==24.2
paginate==0.5.7
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.1.0
pip-tools==7.4.1
platformdirs==4.5.1
pluggy==1.6.0
polars==1.36.1
polars-runtime-32==1.36.1
prometheus-client==0.23.1
prompt-toolkit==3.0.52
protobuf==5.29.5
psutil==7.2.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyjwt==2.10.1
pymdown-extensions==10.20
pyopenssl==25.3.0
pyparsing==3.3.1
pyproject-hooks==1.2.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pyyaml==6.0.3
pyyaml-env-tag==1.1
pyzmq==27.1.0
readme-renderer==44.0
referencing==0.37.0
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
ruff==0.9.3
s3transfer==0.16.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
send2trash==2.0.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==4.1.1
snowflake-snowpark-python[localtest]==1.44.0
sortedcontainers==2.4.0
soupsieve==2.8.1
sparkleframe==0.6.0
sqlalchemy==2.0.45
sqlparse==0.5.5
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
tomli-w==1.2.0
tomlkit==0.13.3
tornado==6.5.4
traitlets==5.14.3
twine==6.1.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.4.0
verspec==0.1.0
watchdog==6.0.0
wcmatch==10.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
werkzeug==3.1.4
wheel==0.45.1
zipp==3.23.0

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
